The problem:
Due to some error when processing a subset of the 20,000 resources before uploading it
into our database, we mistakenly converted all of the descriptions of that subset entirely
into title case (that is, first letter of each word is upper case, the rest of the word in lower
case).

The task:
Write a SQL query or a set of queries, or a script in R, Python, Ruby, or any language of
your choice, that identifies the affected resource site descriptions (= all descriptions that
have been entirely converted into title case), and that provides the descriptions with all
words in the correct casing.

Example:

Your solution might have identified the following description as one that has been
converted by mistake into title case. It reads:
The Foobar Nyc Food Pantry Serves All Populations.
Ideally, your solution converts the description into:
The Foobar NYC food pantry serves all populations.